refactor(hobbies): advanced search form as compound component
HH-371.

Chop old unclear implementation of `AdvancedSearch`-component to a new
compound component, that has the search context provided with React
context tools, form state is initialized with a new hook and the form
section is rendered by a new `AdvancedSearchForm` child component that
uses the new context.
