+++ apps/hobbies-helsinki/src/domain/search/eventSearch/advancedSearch/useAdvancedSearchFormState.tsx
@@ -1,0 +1,105 @@
+import { EVENT_SEARCH_FILTERS } from '@events-helsinki/components';
+import { useRouter } from 'next/router';
+import queryString from 'query-string';
+import React from 'react';
+import { MAPPED_PLACES } from '../constants';
+import { getSearchFilters } from '../utils';
+
+export function useAdvancedSearchFormState() {
+  const router = useRouter();
+  const params: { place?: string } = router.query;
+  const searchParams = React.useMemo(
+    () => new URLSearchParams(queryString.stringify(router.query)),
+    [router.query]
+  );
+  const [categoryInput, setCategoryInput] = React.useState('');
+  const [hobbyTypeInput, setHobbyTypeInput] = React.useState('');
+  const [selectedHobbyTypes, setSelectedHobbyTypes] = React.useState<string[]>(
+    []
+  );
+  const [minAgeInput, setMinAgeInput] = React.useState('');
+  const [maxAgeInput, setMaxAgeInput] = React.useState('');
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [placeInput, setPlaceInput] = React.useState('');
+
+  const [selectedDateTypes, setSelectedDateTypes] = React.useState<string[]>(
+    []
+  );
+  const [selectedCategories, setSelectedCategories] = React.useState<string[]>(
+    []
+  );
+  const [selectedPlaces, setSelectedPlaces] = React.useState<string[]>([]);
+  const [start, setStart] = React.useState<Date | null>(null);
+  const [end, setEnd] = React.useState<Date | null>(null);
+  const [isCustomDate, setIsCustomDate] = React.useState<boolean>(false);
+  const [selectedTexts, setSelectedTexts] = React.useState<string[]>([]);
+  const [autosuggestInput, setAutosuggestInput] = React.useState('');
+
+  // Initialize fields when page is loaded
+  React.useEffect(() => {
+    const {
+      [EVENT_SEARCH_FILTERS.CATEGORIES]: categories,
+      [EVENT_SEARCH_FILTERS.HOBBY_TYPES]: hobbyTypes,
+      [EVENT_SEARCH_FILTERS.DATE_TYPES]: dateTypes,
+      [EVENT_SEARCH_FILTERS.END]: endTime,
+      [EVENT_SEARCH_FILTERS.PLACES]: places,
+      [EVENT_SEARCH_FILTERS.START]: startTime,
+      [EVENT_SEARCH_FILTERS.TEXT]: text,
+      [EVENT_SEARCH_FILTERS.MIN_AGE]: audienceMinAgeLt,
+      [EVENT_SEARCH_FILTERS.MAX_AGE]: audienceMaxAgeGt,
+    } = getSearchFilters(searchParams);
+
+    const pathPlace = params.place && MAPPED_PLACES[params.place.toLowerCase()];
+
+    if (pathPlace) {
+      places.push(pathPlace);
+    }
+
+    setSelectedCategories(categories);
+    setSelectedHobbyTypes(hobbyTypes || []);
+    setSelectedPlaces(places);
+    setSelectedTexts(text || []);
+    setEnd(endTime);
+    setStart(startTime);
+    setMinAgeInput(audienceMinAgeLt || '');
+    setMaxAgeInput(audienceMaxAgeGt || '');
+
+    if (endTime || startTime) {
+      setIsCustomDate(true);
+    } else {
+      setSelectedDateTypes(dateTypes);
+    }
+    setAutosuggestInput(text?.toString() || '');
+  }, [searchParams, params]);
+
+  return {
+    categoryInput,
+    setCategoryInput,
+    hobbyTypeInput,
+    setHobbyTypeInput,
+    selectedHobbyTypes,
+    setSelectedHobbyTypes,
+    minAgeInput,
+    setMinAgeInput,
+    maxAgeInput,
+    setMaxAgeInput,
+    placeInput,
+    setPlaceInput,
+    selectedDateTypes,
+    setSelectedDateTypes,
+    selectedCategories,
+    setSelectedCategories,
+    selectedPlaces,
+    setSelectedPlaces,
+    start,
+    setStart,
+    end,
+    setEnd,
+    isCustomDate,
+    setIsCustomDate,
+    selectedTexts,
+    setSelectedTexts,
+    autosuggestInput,
+    setAutosuggestInput,
+  };
+}
