+++ apps/hobbies-helsinki/src/pages/search/index.tsx
@@ -34,7 +34,7 @@ import getHobbiesStaticProps from '../../domain/app/getHobbiesStaticProps';
 import cmsHelper from '../../domain/app/headlessCmsHelper';
 import { hobbiesApolloClient } from '../../domain/clients/hobbiesApolloClient';
 import serverSideTranslationsWithCommon from '../../domain/i18n/serverSideTranslationsWithCommon';
-import AdvancedSearch from '../../domain/search/eventSearch/AdvancedSearch';
+import AdvancedSearch from '../../domain/search/eventSearch/advancedSearch/AdvancedSearch';
 
 const Search: NextPage<{
   previewToken: string;
